feat: extract cve list and filters from vulnerability exposure
