+++ src/components/security/AddCVEPolicy/VulnerabilityExposure.tsx
@@ -17,9 +17,6 @@
 import { useMemo, useState } from 'react'
 import {
     Progressing,
-    DEFAULT_BASE_PAGE_SIZE,
-    GenericFilterEmptyState,
-    Pagination,
     SearchBar,
     FilterSelectPicker,
     useAsync,
@@ -27,25 +24,23 @@ import {
     SelectPickerOptionType,
     ErrorScreenManager,
     FilterChips,
-    SortableTableHeaderCell,
-    noop,
-    SortingOrder,
     FeatureTitleWithInfo,
     getCVEUrlFromCVEName,
     Button,
     ComponentSizeType,
     ButtonComponentType,
 } from '@devtron-labs/devtron-fe-common-lib'
-import { getCVEControlList, getVulnerabilityFilterData } from './security.service'
+import { getCVEControlList, getVulnerabilityFilterData } from '../security.service'
 import {
     CVEControlListPayload,
-    ExposureListProps,
+    ExposureListContainerProps,
     VulnerabilityExposureFilterKeys,
     VulnerabilityExposureUrlFiltersType,
-} from './security.types'
-import { parseVulnerabilityExposureSearchParams } from './security.util'
+} from '../security.types'
+import { parseVulnerabilityExposureSearchParams } from '../security.util'
+import ExposureList from './ExposureList'
 
-const ExposureList = ({ urlFilters }: ExposureListProps) => {
+const ExposureListContainer = ({ urlFilters }: ExposureListContainerProps) => {
     const [filtersLoading, filtersResponse, filtersError, reloadFilters] = useAsync(getVulnerabilityFilterData)
 
     const {
@@ -190,77 +185,19 @@ const ExposureList = ({ urlFilters }: ExposureListProps) => {
     const appListLength: number = appListResponse.result.totalCount
     const areFiltersApplied: boolean = !!appName || !!cluster.length || !!environment.length
 
-    const renderList = () => {
-        if (!appListLength && areFiltersApplied) {
-            return (
-                <div className="dc__position-rel" style={{ height: 'calc(100vh - 200px)' }}>
-                    <GenericFilterEmptyState handleClearFilters={clearExposureListFilters} />
-                </div>
-            )
-        }
-        return appListLength ? (
-            <>
-                <div className="w-100">
-                    <div className="fs-12 fw-6 lh-20 cn-7 pl-20 pr-20 dc__border-bottom px-20 vulnerability-exp-table-row h-36">
-                        <SortableTableHeaderCell
-                            title="NAME"
-                            isSortable={false}
-                            isSorted
-                            triggerSorting={noop}
-                            disabled={false}
-                            sortOrder={SortingOrder.ASC}
-                        />
-                        <SortableTableHeaderCell
-                            title="ENVIRONMENT"
-                            isSortable={false}
-                            isSorted
-                            triggerSorting={noop}
-                            disabled
-                            sortOrder={SortingOrder.ASC}
-                        />
-                        <SortableTableHeaderCell
-                            title="POLICY"
-                            isSortable={false}
-                            isSorted
-                            triggerSorting={noop}
-                            disabled
-                            sortOrder={SortingOrder.ASC}
-                        />
-                    </div>
-                    {appListResponse.result.scanList.map((cve) => (
-                        <div
-                            key={`${cve.appName}-${cve.envName}`}
-                            className="dc__border-bottom-n1 dc__hover-n50 vulnerability-exp-table-row px-20 h-44 dc__align-items-center fs-13 lh-20 cn-9"
-                        >
-                            <span>{cve.appName}</span>
-                            <span>{cve.envName}</span>
-                            <span>
-                                <span className={`security-tab__cell-policy--${cve.policy}`}>
-                                    {cve.policy.toUpperCase()}
-                                </span>
-                            </span>
-                        </div>
-                    ))}
-                </div>
-                {appListLength > DEFAULT_BASE_PAGE_SIZE && (
-                    <Pagination
-                        rootClassName="flex dc__content-space px-20"
-                        size={appListLength}
-                        pageSize={pageSize}
-                        offset={offset}
-                        changePage={changePage}
-                        changePageSize={changePageSize}
-                    />
-                )}
-            </>
-        ) : null
-    }
-
     return (
         <>
             {!!appListLength && renderFilters()}
             {!!appListLength && renderAppliedFilters()}
-            {renderList()}
+            <ExposureList
+                appListResponse={appListResponse}
+                areFiltersApplied={areFiltersApplied}
+                clearExposureListFilters={clearExposureListFilters}
+                offset={offset}
+                pageSize={pageSize}
+                changePage={changePage}
+                changePageSize={changePageSize}
+            />
         </>
     )
 }
@@ -314,7 +251,7 @@ const VulnerabilityExposure = () => {
                     <Button
                         dataTestId="search-vulnerability-button"
                         text="Search"
-                        size={ComponentSizeType.small}
+                        size={ComponentSizeType.medium}
                         onClick={handleSearchCVEButtonClick}
                         component={ButtonComponentType.button}
                     />
@@ -329,7 +266,7 @@ const VulnerabilityExposure = () => {
                 )}
             </div>
             <div className="dc__border-bottom-n1" />
-            {!!cveName && <ExposureList urlFilters={urlFilters} />}
+            {!!cveName && <ExposureListContainer urlFilters={urlFilters} />}
         </div>
     )
 }
