+++ src/components/security/AddCVEPolicy/index.tsx
@@ -1,0 +1,1 @@
+export { default as VulnerabilityExposure } from './VulnerabilityExposure'
